Add initialization context for mods
Move code from Preprocessors.kt and FilterFactory.kt to the initialization context, and the rest to internal files
Modify EntryPoint class: rename run and add InitializerContext parameter
Modify OptiGlue loading
