+++ OptiGUI/src/main/kotlin/opekope2/optigui/EntryPoint.kt
@@ -1,15 +1,15 @@
 package opekope2.optigui
 
-import  opekope2.optigui.interaction.registerPreprocessor
-import  opekope2.filter.registerFilterFactory
 
 /**
  * OptiGUI entry point. Called when OptiGUI finishes initialization.
- * This is where [registerFilterFactory] and [registerPreprocessor] should be called.
+ * This is where [InitializerContext.registerFilterFactory] and [InitializerContext.registerPreprocessor] should be called.
  */
 fun interface EntryPoint {
     /**
-     * Runs the entry point.
+     * Registers filter factories using [context].
+     *
+     * @param context The initialization context
      */
-    fun run()
+    fun onInitialize(context: InitializerContext)
 }
