+++ OptiGUI/src/main/kotlin/opekope2/optigui/internal/interaction/FilterFactory.kt
@@ -1,0 +1,6 @@
+package opekope2.optigui.internal.interaction
+
+import opekope2.filter.FilterInfo
+import opekope2.optigui.resource.Resource
+
+internal val filterFactories = mutableMapOf<String, MutableSet<(Resource) -> FilterInfo?>>()
