+++ OptiGUI/src/main/kotlin/opekope2/optigui/internal/interaction/Preprocessors.kt
@@ -1,0 +1,9 @@
+package opekope2.optigui.internal.interaction
+
+import net.minecraft.block.entity.BlockEntity
+import net.minecraft.entity.Entity
+
+internal val blockEntityPreprocessors =
+    mutableMapOf<Class<out BlockEntity>, IdentifiableBlockEntityPreprocessor>()
+internal val entityPreprocessors =
+    mutableMapOf<Class<out Entity>, IdentifiableEntityPreprocessor>()
